wip on parsing refacto
