+++ srcs/error.c
@@ -6,7 +6,7 @@
 /*   By: lpassera <lpassera@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/24 14:36:53 by pcharton          #+#    #+#             */
-/*   Updated: 2021/07/12 19:09:25 by pcharton         ###   ########.fr       */
+/*   Updated: 2021/07/14 13:47:01 by pcharton         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,45 +32,21 @@ int	parser_error(t_token *tok)
 	return (-1);
 }
 
-bool	check_first_node_cmd(t_simple_command *cmd)
+void	display_error(char *command, char *custom)
 {
-	t_token	*tok;
+	int		i;
+	char	buffer[256];
+	char	*first;
 
-	tok = cmd->words->content;
-	if (tok && tok->role == operator)
+	ft_bzero(&buffer[0], 256);
+	first = "minishell:  ";
+	i = ft_strlcpy(&buffer[0], first, ft_strlen(first) + 1);
+	i += ft_strlcpy(&buffer[i - 1], command, ft_strlen(command) + 1);
+	if (custom)
 	{
-		parser_error(tok);
-		return (true);
+		i += ft_strlcpy(&buffer[i - 1], ": ", 3);
+		i += ft_strlcpy(& buffer[i - 1], custom, ft_strlen(custom) + 1);
 	}
-	else
-		return (false);
-}
-
-bool	check_syntax_error(t_list *list)
-{
-	t_list	*tmp;
-	t_list	*pipeline_tmp;
-	t_block	*block;
-
-	tmp = list;
-	while (tmp)
-	{
-		block = tmp->content;
-		if ((block->id == simple_command)
-			&& (check_first_node_cmd(block->kind.cmd)))
-			return (true);
-		else if (block->id == pipeline)
-		{
-			pipeline_tmp = block->kind.pipe->commands;
-			while (pipeline_tmp)
-			{
-				if (check_first_node_cmd(pipeline_tmp->content))
-					return (true);
-				pipeline_tmp = pipeline_tmp->next;
-			}
-			return (false);
-		}
-		tmp = tmp->next;
-	}
-	return (false);
+	buffer[i - 1] = '\n';
+	write(2, buffer, ft_strlen(buffer));
 }
