+++ srcs/evaluation.c
@@ -6,7 +6,7 @@
 /*   By: lpassera <lpassera@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/24 14:21:54 by pcharton          #+#    #+#             */
-/*   Updated: 2021/07/12 20:21:30 by pcharton         ###   ########.fr       */
+/*   Updated: 2021/07/14 15:31:54 by pcharton         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -64,33 +64,6 @@ int	flag_simple_command(t_simple_command *list)
 		if (token->flag)
 			remove_mixed_quotes(token);
 		tmp = tmp->next;
-	}
-	return (0);
-}
-
-int	flag_redirection(t_list *list)
-{
-	t_redirection	*redir;
-	t_list			*node;
-	char			*expanded_str;
-
-	node = list;
-	while (node)
-	{
-		redir = node->content;
-		word_flagger(redir->file);
-		expanded_str = expand_text(redir->file->cmd);
-		if (!*expanded_str)
-		{
-			display_error(redir->file->cmd, "Ambiguous redirect");
-			return (1);
-		}
-		else
-		{
-			free(redir->file->cmd);
-			redir->file->cmd = ft_strdup(expanded_str);
-		}
-		node = node->next;
 	}
 	return (0);
 }
