+++ srcs/redirections.c
@@ -6,23 +6,29 @@
 /*   By: lpassera <lpassera@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/13 10:35:34 by lpassera          #+#    #+#             */
-/*   Updated: 2021/07/12 20:22:29 by pcharton         ###   ########.fr       */
+/*   Updated: 2021/07/14 15:43:04 by pcharton         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "header.h"
+
+int	expand_redirection(t_redirection *redirection);
 
 int	handle_redirections(t_list *command)
 {
 	t_redirection	*redirection;
 	t_list			*node;
 
-	if (flag_redirection(command))
-		return (1);
 	node = command;
 	while (node)
 	{
 		redirection = node->content;
+		word_flagger(redirection->file);
+		if (redirection->file->role)
+		{
+			if (expand_redirection(redirection))
+				return (1);
+		}
 		redirection->fd = create_file(redirection->file->cmd,
 				redirection->operator->cmd);
 		if (redirection->fd == -1)
@@ -30,6 +36,29 @@ int	handle_redirections(t_list *command)
 		node = node->next;
 	}
 	return (0);
+}
+
+int	expand_redirection(t_redirection *redirection)
+{
+	char			*expanded_str;
+
+	expanded_str = expand_text(redirection->file->cmd);
+	if (!expanded_str)
+	{
+		ft_exit_with_error_msg(MSG_MALLOC_FAILED);
+		return (-1);
+	}
+	else if (!*expanded_str)
+	{
+		display_error(redirection->file->cmd, "Ambiguous redirect");
+		return (1);
+	}
+	else
+	{
+		free(redirection->file->cmd);
+		redirection->file->cmd = expanded_str;
+		return (0);
+	}
 }
 
 int	create_file(char *path, char *redirection_type)
