+++ srcs/syntax_check.c
@@ -1,0 +1,104 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   syntax_check.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: pcharton <pcharton@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/07/14 13:42:30 by pcharton          #+#    #+#             */
+/*   Updated: 2021/07/14 15:49:14 by pcharton         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "header.h"
+
+bool	check_command_syntax(t_list *list)
+{
+	size_t	index;
+	char	*tmp;
+
+	index = count_command_words(list);
+	if (!index)
+	{
+		if (!ft_strncmp(((t_token *)list->content)->cmd, "|", 1))
+			tmp = "pipe";
+		else if (!ft_strncmp(((t_token *)list->content)->cmd, ";", 1))
+			tmp = "semi-colon";
+		else
+			tmp = "newline";
+		printf("-Minishell: syntax error near unexpected `%s` token\n",
+			tmp);
+		return (false);
+	}
+	else
+		return (true);
+}
+
+bool	check_redirections(t_list *list)
+{
+	t_list			*node;
+	t_redirection	*redir;
+	t_token			*tok;
+
+	node = list;
+	while (node)
+	{
+		redir = node->content;
+		tok = redir->file;
+		if (tok->role != word)
+		{
+			display_error("unexpected syntax error near `newline'", NULL);
+			return (true);
+		}
+		node = node->next;
+	}
+	return (false);
+}
+
+bool	check_first_node_cmd(t_simple_command *cmd)
+{
+	t_token	*tok;
+
+	tok = cmd->words->content;
+	if (tok && tok->role == operator)
+	{
+		parser_error(tok);
+		return (true);
+	}
+	else
+	{
+		if (cmd->redirections)
+			return (check_redirections(cmd->redirections));
+
+	}
+	return (false);
+}
+
+bool	check_syntax_error(t_list *list)
+{
+	t_list	*tmp;
+	t_list	*pipeline_tmp;
+	t_block	*block;
+
+	tmp = list;
+	while (tmp)
+	{
+		block = tmp->content;
+		if ((block->id == simple_command)
+			&& (check_first_node_cmd(block->kind.cmd)))
+			return (true);
+		else if (block->id == pipeline)
+		{
+			pipeline_tmp = block->kind.pipe->commands;
+			while (pipeline_tmp)
+			{
+				if (!pipeline_tmp || check_first_node_cmd(pipeline_tmp->content))
+					return (true);
+				pipeline_tmp = pipeline_tmp->next;
+			}
+			return (false);
+		}
+		tmp = tmp->next;
+	}
+	return (false);
+}
