--- tests/integration/results/output_expansion_error_2
@@ -1,3 +1,0 @@
-free(): invalid pointer
-timeout: la commande surveillée a renvoyé un core dump
-Aborted
