--- tests/integration/results/output_only_redirections
@@ -1,1 +1,0 @@
-Minishell : unexpected syntax error near `|'
