--- tests/integration/results/output_quotes_error
@@ -1,1 +1,0 @@
-"bonjour"'
