--- tests/integration/results/output_quotes_error_3
@@ -1,1 +1,0 @@
-"'"'bonjour
