--- tests/integration/results/output_quotes_error_4
@@ -1,1 +1,0 @@
-bon"\"jour
