--- tests/integration/results/output_quotes_ok_2
@@ -1,1 +1,0 @@
-''bonjour""
