--- tests/integration/results/output_quotes_ok_3
@@ -1,1 +1,0 @@
-bon''jour
